parsers: fixed stylistic errors in create_default_style_elements
