+++ src/parsers/texttracks/webvtt/html/parse_webvtt_to_div.ts
@@ -19,10 +19,10 @@ import getStyleBlocks from "../get_style_blocks";
 import parseCueBlock from "../parse_cue_block";
 import { getFirstLineAfterHeader } from "../utils";
 import convertPayloadToHTML from "./convert_payload_to_html";
+import createDefaultStyleElements from "./create_default_style_elements";
 import parseStyleBlock, {
   IStyleElement,
 } from "./parse_style_block";
-import createDefaultStyleElements from "../create_default_style_elements";
 
 export interface IVTTHTMLCue {
   start : number;
